refactor: Mock 파일 분리
